provider-dialog - separate modal code
